Update Supabase data loader and versioning details
This commit updates the Supabase data loader in different ways. It firstly modifies and expands the external package imports to include more from supabase and tanstack in the vite config. It secondly renames and enhances the comment for the fetch data function and increases the default limit value in the data loader core. The versions of both the packages data-loader-supabase-core and data-loader-supabase-nextjs have been increased to reflect these changes.